{"schema":"decodex/commit/1","summary":"Extract frozen annotation rendering owner","authority":"manual"}
